punto de venta testeada

+++ src/compras/CompraPuntual.java
@@ -1,20 +1,20 @@
 package compras;
 
-import puntosDeVenta.PuntoDeVenta;
+import puntoDeVenta.PuntoDeVenta;
 
 public class CompraPuntual extends Compra {
-	private Integer cantidadDeHoras;
+	private Double cantidadDeHoras;
 
-	public CompraPuntual(PuntoDeVenta puntoDeVenta, Integer cantidadDeHoras) {
+	public CompraPuntual(PuntoDeVenta puntoDeVenta, Double cantidadDeHoras) {
 		super(puntoDeVenta);
 		this.setCantidadDeHoras(cantidadDeHoras);
 	}
 
-	public Integer getCantidadDeHoras() {
+	public Double getCantidadDeHoras() {
 		return cantidadDeHoras;
 	}
 
-	private void setCantidadDeHoras(Integer cantidadDeHoras) {
+	private void setCantidadDeHoras(Double cantidadDeHoras) {
 		this.cantidadDeHoras = cantidadDeHoras;
 	}
 }
